+++ CHANGELOG.md
@@ -5,10 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project might one day adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
-
-
+### User-facing
+- Added "Verbose logging" preference to control debug/info output.  Release builds default to disabled; DEBUG builds default to enabled.
 
 ## [0.0.5] - 2026-02-28
 ### Developer